adds endpoint for placing orders

+++ demo/src/main/java/com/example/demo/repository/item/ItemRepository.java
@@ -7,5 +7,5 @@ public interface ItemRepository{
     Item findByName(String name);
     Item save(Item item);
 
-    Item findById(int id);
+    Item findById(Long id);
 }
